Create A_Litte_Game
This was an exercise from LearnPythonTheHardway.org, I think it was exercise 36 where we were to create a little game performing if and while loops as well as calling functions inside other functions. I think the overall goal here is to have fun and be creative, mainly because I tend to sometimes get bored.